Misc refactor
