--- packages/next-sitemap/src/dynamic/index.ts
@@ -1,3 +1,0 @@
-export * from './sitemap'
-export * from './sitemap-index'
-export * from './response'
