+++ packages/next-sitemap/src/index.ts
@@ -1,4 +1,12 @@
 export * from './sitemap/build'
-export * from './sitemap-index/build'
-export * from './dynamic'
-export * from './interface'
+
+// Export types
+export * from './interface.js'
+
+// Export server side sitemaps
+export * from './ssr/response.js'
+export * from './ssr/sitemap-index.js'
+export * from './ssr/sitemap.js'
+
+// Export builders
+export * from './builder/sitemap-index.js'
