+++ packages/next-sitemap/src/ssr/response.ts
