+++ packages/next-sitemap/src/ssr/sitemap.ts
@@ -1,8 +1,14 @@
-import type { ISitemapField } from '../interface'
-import { buildSitemapXml } from '../sitemap/build'
 import type { GetServerSidePropsContext } from 'next'
-import { withXMLResponse } from './response'
+import { withXMLResponse } from './response.js'
+import { buildSitemapXml } from '../builder/sitemap.js'
+import type { ISitemapField } from '../interface.js'
 
+/**
+ * Generate server side sitemaps
+ * @param ctx
+ * @param fields
+ * @returns
+ */
 export const getServerSideSitemap = async (
   ctx: GetServerSidePropsContext,
   fields: ISitemapField[]
